requirements.txt
Flask==2.3.2
gspread==5.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.93.0
functions-framework==3.8.1
gunicorn==22.0.0

